Review comment updates
