--- deploy/0_preconditions.ts
@@ -1,14 +1,0 @@
-import { Logger } from 'ethers/lib/utils';
-import { validateDeployment, configureDeploymentSettings } from '../utils/deploy';
-import { LogLevel } from '@ethersproject/logger';
-
-export const deploy: CustomHardhatDeployFunction = async (env) => {
-    const hre = env as unknown as CustomHardHatRuntimeEnvironment;
-    Logger.setLogLevel(LogLevel.DEBUG);
-    hre.log(`0_preconditions`);
-    validateDeployment({ hre });
-    await configureDeploymentSettings({ hre });
-};
-
-export default deploy;
-deploy.tags = ['preconditions'];
