--- deploy/1_seed.ts
@@ -1,16 +1,0 @@
-import { Logger } from 'ethers/lib/utils';
-import { validateDeployment, configureDeploymentSettings, seedContracts } from '../utils/deploy';
-import { LogLevel } from '@ethersproject/logger';
-import { getContractsFromDeployments } from '../test/helpers/index';
-
-export const deploy: CustomHardhatDeployFunction = async (env) => {
-    const hre = env as unknown as CustomHardHatRuntimeEnvironment;
-    Logger.setLogLevel(LogLevel.DEBUG);
-    hre.log(`1_seed`);
-    const contracts = await getContractsFromDeployments(hre);
-    await seedContracts({ hre, contracts });
-};
-
-export default deploy;
-deploy.tags = ['seed'];
-deploy.runAtTheEnd = true;
