+++ utils/contracts.ts
@@ -1,4 +1,5 @@
 import { Contract } from 'ethers';
+import { Contracts } from '@/utils/deploy';
 import type {
   BridgedPolygonNORI,
   Certificate,
@@ -7,6 +8,31 @@ import type {
   NORI,
   Removal,
 } from '../typechain-types';
+
+export const getContractsFromDeployments = async (
+    hre: CustomHardHatRuntimeEnvironment
+  ): Promise<Required<Contracts>> => {
+    const deployments = await hre.deployments.all();
+    const contracts = {
+      NORI: deployments['NORI']?.address ? await getNORI({ hre }) : undefined,
+      BridgedPolygonNORI: deployments['BridgedPolygonNORI']?.address
+        ? await getBridgedPolygonNori({ hre })
+        : undefined,
+      LockedNORI: deployments['LockedNORI']?.address
+        ? await getLockedNORI({ hre })
+        : undefined,
+      FIFOMarket: deployments['FIFOMarket']?.address
+        ? await getFIFOMarket({ hre })
+        : undefined,
+      Removal: deployments['Removal']?.address
+        ? await getRemoval({ hre })
+        : undefined,
+      Certificate: deployments['Certificate']?.address
+        ? await getCertificate({ hre })
+        : undefined,
+    } as Required<Contracts>;
+    return contracts;
+  };
 
 export const getContract = async ({
   contractName,
